Updated the default config file to take into account changes in the code.
Fixed the comments in the header.

Bumped to v1.1.

Moved the "from" imports out of the "import"s.

Replaced optparse with argparse.  It simplified the code quite a bit.  Got
rid of the get_options() function at the same time.

Knocked two TODOs off the list.

Moved global variables and handles up to the global variable bit to make the
code a bit easier to read weeks or months later.

Reworked logging slightly to be more in line with how the logging module is
meant to be used.

Reformatted the core code to be one Python indent (four spaces) closer to the
left.  It's more my style.

Signed-off-by: The Doctor <drwho@virtadpt.net>