Started stripping out some debug logging that really isn't necessary now that the core code is stable.
Stripped out the "run a manual search" block of code because it isn't
necessary anymore.  This bot is going to run in the background on an
autonomous basis.  You'll undoubtedly hear me swearing loudly if I have to
put it all back at some future time.

Added an extra check to ensure that both the number of requested searches
and the length of the search result list are zero, otherwise the bot
makes empty search requests.

Added a check at the top of the "do stuff" loop to see if the darknet bot
name has been filled in from the config file.  If it hasn't, it gets skipped,
else the list makes an HTTP request to /, gets the "no commands" response,
and triggers on that.

Signed-off-by: The Doctor <drwho@virtadpt.net>