Bumped web_search_bot.py to v1.1.
Added a little helpful output to web_search_bot.py so I know if a search
result has been successfully picked out of the message queue if I happen to
be logged into the server.

Wrapped polling the message queue in a try..except block because, as it was
before, if the message queue was offline or too many polling attempts timed
out, the requests module would kill the web search agent.

Added code to handle the "there are no search results" corner case.

Signed-off-by: The Doctor <drwho@virtadpt.net>