Added configuration options to web_search_bot.conf which set up transmission of search results via e-mail by the bot.
Added comments to the v1.1 part, deleted some of the TODO comments.

Pulled in a few more modules to facilitate e-mailing search results.

Added numbers (up to thirty) to be used by the command parser.

Added functionality in which the bot can contact an SMTP server (specifically,
one which implicitly trusts the machine this bot is running on) and send an
e-mail containing search results.

Added code to the command parser which figures out that the user wants to
e-mail search results.

Updated logging functinoality.

Signed-off-by: The Doctor <drwho@virtadpt.net>