Removed the \# from the channel name because I don't even.
Temporarily uncommented the 'brain' option in the configuration file because
I need it to test the bot as I build it.  This will probably change in the
future.

Started pulling in the irc submodules as I need them.

Started working on the actual bot class.

Changed the default nickname the bot will try to take.

Reworked the code that tries to open and load the config file.