Reworked the header comments a little.
Updated the TODO list.

Pulled in the floating point division module.

Moved most of the constants inside the DixieBot() class because they're
technically required in there only.

Because the docs for the irc module are lacking, I've been poking around in
the objects' respective directories and taking notes.

Rewrote DixieBot.__init__() slightly.

The callback methods look for events, not exceptions.  The prototypes for
irc.bot.SingleServerIRCBot() look like this:

    def on_something(self, connection, event):

Added callbacks on_privmsg() and on_pubmsg(), which does most of the heavy
lifting for the bot's intended purpose.

Added proper logging.

Chopped out the last of the legacy code.