Initial commit of requirements.txt.
Changed the ratio of long words-to-words to a static attribute on the
DixieBot class.  Wondering if I should make this percentage configurable
(in the config file and on the command line), and maybe if it should be
changeable on the fly by the bot's owner (say, with a private message).

Updated TODO list a little.

Signed-off-by: The Doctor <drwho@virtadpt.net>