Updated the README file with some new information.
Added Cobe to requirements.txt.

Stripped MegaHAL out of the codebase and ported everything over to Cobe.

Removed the ability to change the order of the Markov chains because there
really isn't any need to.

Removed the static class attributes and line complexity stuff from DixieBot
because it's not really needed for anything anymore.  Cobe does all that and
more.

Accidentally left a debugging print in DixieBot.on_pubmsg().  Oops.  I'll
get rid of it later.

Wrote a simple parser to figure out if the construct is being addressed
directly by it's owner ("bot: Hello sailor!") or if its owner is just
speaking in the channel.  Either way it'll always learn from its owner's
text but in the former case it'll always respond; in the latter case it has
a 10% probability of responding.

Was accidentally training the Markov brain on raw IRC objects instead of
text posted to channels.  Oops.

Updated some of the online help text.

Fixed a subtle bug in which the bot would always ignore the path to an
existing brainfile specified in the config file and require that a path be
passed as an argument.

Pushed configuration of the logger farther up in the setup process so that
proper logging could be used for user output.

Fixed bugs in which an IRC server and channel specified in the config file
only would be ignored and had to be passed as command line arguments.

Added a little informational output during setup so the user could double
check.

Rewrote the Markov brain's initial training sequence.  I had to because I
started using a different implementation.