Added password authentication support to the bot.  Nicks on IRC are inherently changeable so a password can be either put into the config file or passed on the command line.  This also sets the nick the bot considers its owner, in case your current nick is taken by someone else.
Added authentication checks to the bot that don't let the "admin only" bits
trigger unless the person

Added a little code that'll greet the channel 10% of the time.  Because why
not.

Added a couple of !commands.  Use !help or !commands to list them.

Reworked how the Markov brain is detected and trained on the command line.
This also, incidentally, makes it possible to train the brain additional times
from text files.