Added stubs for IRC, Slack, and XMPP bots to beta_fork.
Minor change to exocortex_xmpp_bridge.py after profiling it for a while and
watching it lock up.  Evidence suggests that a network socket underlying it
(possible the XMPP server) went away and it didn't know how to handle it.
Prosody is restarted every couple of days to rotate its logs so this might
account for it.

Signed-off-by: The Doctor <drwho@virtadpt.net>