Added instructions for e-mailing search results to exocortex_xmpp_bridge.py.
Pulled 'zero' out of the hash of acceptible numbers of search terms in
web_search_bot.py because that edge case is caught elsewhere in the code.

Explicitly declared the words[] array in parse_search_request() so that we
can be absolutely certain that it's empty on each evocation.

Added some synonyms for "e-mail" to the command parser in parse_search_request.

Fixed the code that takes stuff out of the parsed command in the
parse_search_request function of web_search_bot.py.  It really didn't make
much sense.

Explicitly blanked the destination_email_address variable at the bottom of
the "e-mail search results to this address" code in web_search_bot.py to fix
the "Why did it only work once?" bug.

Signed-off-by: The Doctor <drwho@virtadpt.net>