Updated the README somewhat.
Fixed a bug in which I was testing for something returned by the
Etherpad-Lite object, when in fact it doesn't return anything.  Weird, but
there you have it.

Fixed a bug in the default config file in which stuff was in quotes that
shouldn't be in quotes.  The Etherpad-Lite client was being literal with
them, i.e., quotes and all, and that was causing it to error out.  Oops.

Signed-off-by: The Doctor <drwho@virtadpt.net>