Removed a call to .lower() from the XMPP bridge which I think is breaking some other stuff that relies upon it (i.e., URLs).
Replaced the SHA-1 hash of an downloaded article by paywall breaker with
the title because that was just a dumb idea.  I can't find anything in the
Etherpad-Lite index that way.

Added three modules to paywall breaker's requirements.txt file because
by default requests (which sits on top of urllib3) isn't SNI aware and that
can cause some stuff to go sideways.

Signed-off-by: The Doctor <drwho@virtadpt.net>