Pulled out the Unicode module because it wasn't useful for anything.
Went back to using SHA-1 hashes of the downloaded and parsed page because I
got tired of fighting with the padID part of the Etherpad-Lite API and not
being able to use Paywall Breaker in the interim.  I'm starting to suspect
that it has to do with the maximum length of the padID.  Either that, or it's
been way too long since I've used anybody's pad but my own.

Signed-off-by: The Doctor <drwho@virtadpt.net>