I had the required_keys[] list for do_PUT() in the wrong place.  Oops.
Switched around capitalization for the Content-Type HTTP header a bit.
Upon reflection, I didn't have to do this because I normalize them anyway.
This is why I shouldn't be coding after two days of no sleep and with
jet lag.

Fixed a big where a variable name from another of my constructs isn't
used in the XMPP bridge construct.  I really need to break all of this
stuff out into a module so it can be included.  It'll keep this from
happening again and again and again...

Fixed the bug in message_queues['replies'] - it's a list and not a Unicode
string, so I needed to .append() it and not = it.

Pulled in the _send_http_response() and _read_content() methods from the
paywall breaker construct because they're useful, known to work, and are
self contained.  Again with pulling this stuff into a module.

Added a little code to paywall_breaker.py make it easier to set up the
URL to its replies rail.

Added some basic help text to paywall_breaker.py, which can be accessed by
the user typing "help".

Signed-off-by: The Doctor <drwho@virtadpt.net>