exocortex_xmpp_bridge.py's do_PUT() method needed to send an HTTP response code to work properly.
In paywall_breaker.py, changed the declaration of the queue to message_queue
because that's what it's called everywhere else in the code.  Also added a
string 'server' which holds only the server part of the URL (i.e.,
http://server:port/).

Installed a check for an online help request.  Oops.

Reworked online help a bit.  Declared help_reply before trying to reference
it.  Finally got the "PUT the reply to /replies" bit working after a couple
of hours of sleep.

Signed-off-by: The Doctor <drwho@virtadpt.net>