Added to the ever-growing to-do list for Paywall Breaker.
Added a bit of logging that says when the user asks for online help when
in debug mode.

Changed the e-mail sent when the URL is invalid to a message back to the
user via the XMPP bridge.

Changed the "it didn't work" message to be a little more descriptive.

Fixed a couple of glitches in which the wrong XMPP message back to the user
was sent.

Added a "can't send an e-mail" message which goes over the XMPP bridge.

Ran tests in which plain text was grabbed by Paywall Breaker.  Things look
pretty stable at the moment.

Signed-off-by: The Doctor <drwho@virtadpt.net>