Updated the online help for paywall_breaker.py.
Added an online help function to web_index_bot.py.  This involved adding a
check to the command parser, transplanting send_message_to_user() into the
code, and adding the proper response to a help request to the main loop
(which I really should move into a function by itself).

Added a response to the user for the event that something goes wrong.

Signed-off-by: The Doctor <drwho@virtadpt.net>