Fixed a bug in which a truncated search command will cause web_search_bot.py to crash.
Now, if you tell the bot to e-mail its results but don't give it an e-mail
address (or give it one that doesn't match the check) it'll use the default
address.  Which is why I gave it a default address to begin with.  Added
that to the help output, by the way.

Signed-off-by: The Doctor <drwho@virtadpt.net>