Took a "back" out of the response in the XMPP bridge.
Added PyParsing to system_bot.py.

Started defining an interactive command parser so system_bot.py can do some
things on behalf of the user in addition to active monitoring.

Deleted some debugging output that wound up making it more difficult to
figure out what was going on.

Added online help, system load, system info, and available CPUs commands to
the bot.  Available CPUs doesn't work right yet.

Worked on the online help.

Pulled the default value for polling_time out of the argparser and put it at
the top of the utility because it meant the value in the config file (if found)
was getting ignored.  I really need to stop doing that.

Changed the pre-startup debugging output a bit.

Added system info and CPUs (which doesn't work yet) to the list of interactive
commands.

For the record, I had to explicitly cast the variables in the "should I poll
the message queue or not?" conditional to integers because trying to compare
a floating point value to an integer messed with me for a couple of hours
tonight.

Signed-off-by: The Doctor <drwho@virtadpt.net>