Updated README.md file, because it was many revisions out of date.
Initial commit of the exocortex_sip_client.conf.example file, which obviates
the need to edit the exocortex_sip_client.py file to configure it (which is
as bad as it gets).

Initial commit of what should probably be v2.0 of exocortex_sip_client.py,
given how much I've reworked and added to it.

Signed-off-by: The Doctor <drwho@virtadpt.net>