Forgot that I could add a description in the online help text of exocortex_sip_client.py.  Added that.
Added an entry to the default config file to specify the SIP client.

Pretty much rewrote exocortex_web_to_speech.py, so call this v2.0.  Took
out web.py and replaced it with BaseHTTPServer, changed a bunch of comments,
cleaned up the code some, updated online docs... transplanted a lot of the
helper code I wrote for the XMPP bridge in.

Signed-off-by: The Doctor <drwho@virtadpt.net>