Added a basic corpus for asking help with commands.  Also added this online help.
Fixed a bug where the minimum match confidence was being interpreted as a
string and not an integer from the config file.

Fixed the "I'm not sure" reply a bit.

Added conditional short-circuiting to the command parser to make it a bit
more efficient.  This is probably a bad idea.

Signed-off-by: The Doctor <drwho@virtadpt.net>