Added help corpora for audio and video playback.  Added the online help for same, also.
Updated the basic help a bit.

Added the humanfriendly module to the requirements and code because it does
the heavy lifting of formatting things for people.

Added a function that tries to load a local dump of the media library database
if it exists.  This means that I also added code to save a local dump as a
JSON document after loading the media library from Kodi.  I need to turn that
part into its own function later.

Signed-off-by: The Doctor <drwho@virtadpt.net>