Added more stuff to the search_requests_albums.txt corpus.
Declared a few more variables to document them.

Wrote some code that searches the music library for an album title.  It's
a work in progress, it doesn't quite work yet.  Plus, it's going to need to
be refactored because the main loop will turn into a hairball if I don't.

Added a generic search-the-media-library function to kodi_library.py.  It's
not as accurate as I'd like it to be and is going to need some tinkering.  I
suspect that my earlier experiments were a little too simplistic.

Signed-off-by: The Doctor <drwho@virtadpt.net>