Added a minimum confidence metric for proper matches - titles, names, things that require a much greater confidence than just figuring out what the user means.
Added a function that subtracts the part of the corpus that best matched
the user's text from the user's input and returns it as a string.  This
decreases the amount of text the fuzzy matcher has to go on and improves the
accuracy of results.

Changed up how the main do-stuff loop interprets result metrics a bit by
handling the no-case: No matches at all (confidence: 0).

Added searching by genre.

Added searching by song title.

Signed-off-by: The Doctor <drwho@virtadpt.net>