Fixed a bug where I forgot the match_confidence.  Oops.
Added search of the Kodi box's music library (distinct from the song library)
to the song search.  This is probably a bad idea because it would make sense
to only search one or the other (depending on which one exists or doesn't at
the time) but I haven't tested it enough to know for sure.  Watch out for
this code going away in the future.

Signed-off-by: The Doctor <drwho@virtadpt.net>