Kodibot:
Changed some of the responses to the user.  Changed some of the string
cleaning code.  Added "what's playing?" capability.  Added "play by filename"
and "play by song ID" capability.  These commits aren't ready to go yet
because I really wanted to get the following committed...

Systembot:

Discovered that Systembot wasn't actually sending warning when it was
supposed to because the send_message_to_user() function wasn't callable.
Added it as a parameter to the functions that needed it in the system_stats.py
module.

Signed-off-by: The Doctor <drwho@virtadpt.net>