Edited the template bot to make it a little more internally consistent. You could tell that I based it off of the Web Index Bot at one point.
Initial commit of a run.sh script.

Updated the example config file a bit.

Fleshed out more of the bot itself.  I can't test it because Windbringer
doesn't have the RasPi's GPIO subsystem, but pushing to Github also makes
it easier to pull my code to the test unit for debugging.

Signed-off-by: The Doctor <drwho@virtadpt.net>