Added some fraction rounding to the anemometer and BME280 data samples. Because staring at 10 decimal digits is so easy to debug.
Added an example config file.  That'll change in the future but I don't want
to forget it.

Working on the bot's core.  It's totally not done, I've got a bunch of MOOFs
all through it.  Do not expect this to work because I'm still figuring things
out (unless you want to file any pull requests - that would be awesome).

Signed-off-by: The Doctor <drwho@virtadpt.net>