Started factoring conversions out into a library module.
Getting rid of data elements that are themselves conversions because I'm
setting up the rest of the code to let the user choose metric or
imperial.  I'm eventually going to start adding modules to send data to
citizen science projects, so it seems a good idea to have flexibility to
configure the measurement units they're expecting.

Signed-off-by: The Doctor <drwho@virtadpt.net>